pyproject.toml
[tool.poetry]
name = "django-office-facts"
version = "0.1.0"
description = ""
authors = ["josephlee3454 <joseph.lee3454@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.11.1"
django = "^3.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
